+++ tests/src/Kernel/Cron/CronPushTest.php
@@ -107,7 +107,7 @@ class CronPushTest extends KernelTestBase {
    * @return array
    *   Return an array of arrays.
    */
-  public function goodIntervalProvider() {
+  public static function goodIntervalProvider(): iterable {
     return [
       [
         'en', time() - 4000, 3600,
@@ -148,7 +148,7 @@ class CronPushTest extends KernelTestBase {
    * @return array
    *   Return an array of arrays.
    */
-  public function badIntervalProvider() {
+  public static function badIntervalProvider(): iterable {
     return [
       [
         'en', time() - 4000, 50000,